+++ fractal_server/app/models/v2/history.py
@@ -18,16 +18,21 @@ class HistoryRun(SQLModel, table=True):
     model_config = ConfigDict(arbitrary_types_allowed=True)
 
     id: Optional[int] = Field(default=None, primary_key=True)
-    dataset_id: int = Field(foreign_key="datasetv2.id")
+    dataset_id: int = Field(
+        foreign_key="datasetv2.id",
+        ondelete="CASCADE",
+    )
     workflowtask_id: Optional[int] = Field(
-        foreign_key="workflowtaskv2.id", default=None
+        foreign_key="workflowtaskv2.id",
+        default=None,
+        ondelete="SET NULL",
     )
 
     workflowtask_dump: dict[str, Any] = Field(
-        sa_column=Column(JSONB, nullable=False)
+        sa_column=Column(JSONB, nullable=False),
     )
     task_group_dump: dict[str, Any] = Field(
-        sa_column=Column(JSONB, nullable=False)
+        sa_column=Column(JSONB, nullable=False),
     )
 
     timestamp_started: datetime = Field(
@@ -39,23 +44,34 @@ class HistoryRun(SQLModel, table=True):
 
 
 class HistoryUnit(SQLModel, table=True):
-
     id: Optional[int] = Field(default=None, primary_key=True)
-    history_run_id: int = Field(foreign_key="historyrun.id")
+    history_run_id: int = Field(
+        foreign_key="historyrun.id",
+        ondelete="CASCADE",
+    )
 
     logfile: Optional[str]
     status: str
     zarr_urls: list[str] = Field(
-        sa_column=Column(ARRAY(String)), default_factory=list
+        sa_column=Column(ARRAY(String)),
+        default_factory=list,
     )
 
 
 class HistoryImageCache(SQLModel, table=True):
-
     zarr_url: str = Field(primary_key=True)
-    dataset_id: int = Field(primary_key=True, foreign_key="datasetv2.id")
+    dataset_id: int = Field(
+        primary_key=True,
+        foreign_key="datasetv2.id",
+        ondelete="CASCADE",
+    )
     workflowtask_id: int = Field(
-        primary_key=True, foreign_key="workflowtaskv2.id"
+        primary_key=True,
+        foreign_key="workflowtaskv2.id",
+        ondelete="CASCADE",
     )
 
-    latest_history_unit_id: int = Field(foreign_key="historyunit.id")
+    latest_history_unit_id: int = Field(
+        foreign_key="historyunit.id",
+        ondelete="CASCADE",
+    )