+++ fractal_server/app/runner/v2/runner_functions.py
@@ -6,6 +6,7 @@ from typing import Literal
 from typing import Optional
 
 from pydantic import ValidationError
+from sqlmodel import update
 
 from ..exceptions import JobExecutionError
 from .deduplicate_list import deduplicate_list
@@ -13,6 +14,10 @@ from .merge_outputs import merge_outputs
 from .runner_functions_low_level import run_single_task
 from .task_interface import InitTaskOutput
 from .task_interface import TaskOutput
+from fractal_server.app.db import get_sync_db
+from fractal_server.app.history.status_enum import XXXStatus
+from fractal_server.app.models.v2 import HistoryImageCache
+from fractal_server.app.models.v2 import HistoryUnit
 from fractal_server.app.models.v2 import TaskV2
 from fractal_server.app.models.v2 import WorkflowTaskV2
 from fractal_server.app.runner.components import _COMPONENT_KEY_
@@ -88,7 +93,8 @@ def run_v2_task_non_parallel(
     workflow_dir_remote: Optional[Path] = None,
     executor: BaseRunner,
     submit_setup_call: callable = no_op_submit_setup_call,
-    history_item_id: int,
+    dataset_id: int,
+    history_run_id: int,
 ) -> tuple[TaskOutput, int, dict[int, BaseException]]:
     """
     This runs server-side (see `executor` argument)
@@ -115,6 +121,29 @@ def run_v2_task_non_parallel(
     )
     function_kwargs[_COMPONENT_KEY_] = _index_to_component(0)
 
+    # Database History operations
+    with next(get_sync_db()) as db:
+        history_unit = HistoryUnit(
+            history_run_id=history_run_id,
+            status=XXXStatus.SUBMITTED,
+            logfile=None,  # FIXME
+            zarr_urls=function_kwargs["zarr_urls"],
+        )
+        db.add(history_unit)
+        db.commit()
+        db.refresh(history_unit)
+        history_unit_id = history_unit.id
+        for zarr_url in function_kwargs["zarr_urls"]:
+            db.merge(
+                HistoryImageCache(
+                    workflowtask_id=wftask.id,
+                    dataset_id=dataset_id,
+                    zarr_url=zarr_url,
+                    latest_history_unit_id=history_unit_id,
+                )
+            )
+        db.commit()
+
     result, exception = executor.submit(
         functools.partial(
             run_single_task,
@@ -124,18 +153,30 @@ def run_v2_task_non_parallel(
             root_dir_remote=workflow_dir_remote,
         ),
         parameters=function_kwargs,
-        history_item_id=history_item_id,
         **executor_options,
     )
 
     num_tasks = 1
-    if exception is None:
-        if result is None:
-            return (TaskOutput(), num_tasks, {})
+    with next(get_sync_db()) as db:
+        if exception is None:
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.DONE)
+            )
+            db.commit()
+            if result is None:
+                return (TaskOutput(), num_tasks, {})
+            else:
+                return (_cast_and_validate_TaskOutput(result), num_tasks, {})
         else:
-            return (_cast_and_validate_TaskOutput(result), num_tasks, {})
-    else:
-        return (TaskOutput(), num_tasks, {0: exception})
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.FAILED)
+            )
+            db.commit()
+            return (TaskOutput(), num_tasks, {0: exception})
 
 
 def run_v2_task_parallel(
@@ -147,10 +188,12 @@ def run_v2_task_parallel(
     workflow_dir_local: Path,
     workflow_dir_remote: Optional[Path] = None,
     submit_setup_call: callable = no_op_submit_setup_call,
-    history_item_id: int,
+    dataset_id: int,
+    history_run_id: int,
 ) -> tuple[TaskOutput, int, dict[int, BaseException]]:
 
     if len(images) == 0:
+        # FIXME: Do something with history units/images?
         return (TaskOutput(), 0, {})
 
     _check_parallelization_list_size(images)
@@ -163,14 +206,38 @@ def run_v2_task_parallel(
     )
 
     list_function_kwargs = []
-    for ind, image in enumerate(images):
-        list_function_kwargs.append(
-            dict(
-                zarr_url=image["zarr_url"],
-                **(wftask.args_parallel or {}),
-            ),
-        )
-        list_function_kwargs[-1][_COMPONENT_KEY_] = _index_to_component(ind)
+    history_unit_ids = []
+    with next(get_sync_db()) as db:
+        for ind, image in enumerate(images):
+            list_function_kwargs.append(
+                dict(
+                    zarr_url=image["zarr_url"],
+                    **(wftask.args_parallel or {}),
+                ),
+            )
+            list_function_kwargs[-1][_COMPONENT_KEY_] = _index_to_component(
+                ind
+            )
+            history_unit = HistoryUnit(
+                history_run_id=history_run_id,
+                status=XXXStatus.SUBMITTED,
+                logfile=None,  # FIXME
+                zarr_urls=[image["zarr_url"]],
+            )
+            # FIXME: this should be a bulk operation
+            db.add(history_unit)
+            db.commit()
+            db.refresh(history_unit)
+            db.merge(
+                HistoryImageCache(
+                    workflowtask_id=wftask.id,
+                    dataset_id=dataset_id,
+                    zarr_url=image["zarr_url"],
+                    latest_history_unit_id=history_unit.id,
+                )
+            )
+            db.commit()
+            history_unit_ids.append(history_unit.id)
 
     results, exceptions = executor.multisubmit(
         functools.partial(
@@ -181,11 +248,12 @@ def run_v2_task_parallel(
             root_dir_remote=workflow_dir_remote,
         ),
         list_parameters=list_function_kwargs,
-        history_item_id=history_item_id,
         **executor_options,
     )
 
     outputs = []
+    history_unit_ids_done: list[int] = []
+    history_unit_ids_failed: list[int] = []
     for ind in range(len(list_function_kwargs)):
         if ind in results.keys():
             result = results[ind]
@@ -194,10 +262,25 @@ def run_v2_task_parallel(
             else:
                 output = _cast_and_validate_TaskOutput(result)
             outputs.append(output)
+            history_unit_ids_done.append(history_unit_ids[ind])
         elif ind in exceptions.keys():
             print(f"Bad: {exceptions[ind]}")
+            history_unit_ids_failed.append(history_unit_ids[ind])
         else:
             print("VERY BAD - should have not reached this point")
+
+    with next(get_sync_db()) as db:
+        db.execute(
+            update(HistoryUnit)
+            .where(HistoryUnit.id.in_(history_unit_ids_done))
+            .values(status=XXXStatus.DONE)
+        )
+        db.execute(
+            update(HistoryUnit)
+            .where(HistoryUnit.id.in_(history_unit_ids_failed))
+            .values(status=XXXStatus.FAILED)
+        )
+        db.commit()
 
     num_tasks = len(images)
     merged_output = merge_outputs(outputs)
@@ -214,7 +297,8 @@ def run_v2_task_compound(
     workflow_dir_local: Path,
     workflow_dir_remote: Optional[Path] = None,
     submit_setup_call: callable = no_op_submit_setup_call,
-    history_item_id: int,
+    dataset_id: int,
+    history_run_id: int,
 ) -> tuple[TaskOutput, int, dict[int, BaseException]]:
 
     executor_options_init = submit_setup_call(
@@ -237,6 +321,33 @@ def run_v2_task_compound(
         **(wftask.args_non_parallel or {}),
     )
     function_kwargs[_COMPONENT_KEY_] = f"init_{_index_to_component(0)}"
+
+    # Create database History entries
+    input_image_zarr_urls = function_kwargs["zarr_urls"]
+    with next(get_sync_db()) as db:
+        # Create a single `HistoryUnit` for the whole compound task
+        history_unit = HistoryUnit(
+            history_run_id=history_run_id,
+            status=XXXStatus.SUBMITTED,
+            logfile=None,  # FIXME
+            zarr_urls=input_image_zarr_urls,
+        )
+        db.add(history_unit)
+        db.commit()
+        db.refresh(history_unit)
+        history_unit_id = history_unit.id
+        # Create one `HistoryImageCache` for each input image
+        for zarr_url in input_image_zarr_urls:
+            db.merge(
+                HistoryImageCache(
+                    workflowtask_id=wftask.id,
+                    dataset_id=dataset_id,
+                    zarr_url=zarr_url,
+                    latest_history_unit_id=history_unit_id,
+                )
+            )
+        db.commit()
+
     result, exception = executor.submit(
         functools.partial(
             run_single_task,
@@ -246,8 +357,6 @@ def run_v2_task_compound(
             root_dir_remote=workflow_dir_remote,
         ),
         parameters=function_kwargs,
-        history_item_id=history_item_id,
-        in_compound_task=True,
         **executor_options_init,
     )
 
@@ -258,6 +367,13 @@ def run_v2_task_compound(
         else:
             init_task_output = _cast_and_validate_InitTaskOutput(result)
     else:
+        with next(get_sync_db()) as db:
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.FAILED)
+            )
+            db.commit()
         return (TaskOutput(), num_tasks, {0: exception})
 
     parallelization_list = init_task_output.parallelization_list
@@ -269,6 +385,13 @@ def run_v2_task_compound(
     _check_parallelization_list_size(parallelization_list)
 
     if len(parallelization_list) == 0:
+        with next(get_sync_db()) as db:
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.DONE)
+            )
+            db.commit()
         return (TaskOutput(), 0, {})
 
     list_function_kwargs = []
@@ -293,12 +416,12 @@ def run_v2_task_compound(
             root_dir_remote=workflow_dir_remote,
         ),
         list_parameters=list_function_kwargs,
-        history_item_id=history_item_id,
         in_compound_task=True,
         **executor_options_compute,
     )
 
     outputs = []
+    failure = False
     for ind in range(len(list_function_kwargs)):
         if ind in results.keys():
             result = results[ind]
@@ -307,8 +430,27 @@ def run_v2_task_compound(
             else:
                 output = _cast_and_validate_TaskOutput(result)
             outputs.append(output)
+
         elif ind in exceptions.keys():
             print(f"Bad: {exceptions[ind]}")
+            failure = True
+        else:
+            print("VERY BAD - should have not reached this point")
+
+    with next(get_sync_db()) as db:
+        if failure:
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.FAILED)
+            )
+        else:
+            db.execute(
+                update(HistoryUnit)
+                .where(HistoryUnit.id == history_unit_id)
+                .values(status=XXXStatus.DONE)
+            )
+        db.commit()
 
     merged_output = merge_outputs(outputs)
     return (merged_output, num_tasks, exceptions)