+++ fractal_server/migrations/versions/fbce16ff4e47_new_history_items.py
@@ -1,8 +1,8 @@
-"""Add new history tables
+"""new history items
 
-Revision ID: 5f84657878a9
+Revision ID: fbce16ff4e47
 Revises: af1ef1c83c9b
-Create Date: 2025-03-14 11:49:33.356108
+Create Date: 2025-03-14 15:25:01.083619
 
 """
 import sqlalchemy as sa
@@ -11,7 +11,7 @@ from alembic import op
 from sqlalchemy.dialects import postgresql
 
 # revision identifiers, used by Alembic.
-revision = "5f84657878a9"
+revision = "fbce16ff4e47"
 down_revision = "af1ef1c83c9b"
 branch_labels = None
 depends_on = None
@@ -45,11 +45,13 @@ def upgrade() -> None:
             ["dataset_id"],
             ["datasetv2.id"],
             name=op.f("fk_historyrun_dataset_id_datasetv2"),
+            ondelete="CASCADE",
         ),
         sa.ForeignKeyConstraint(
             ["workflowtask_id"],
             ["workflowtaskv2.id"],
             name=op.f("fk_historyrun_workflowtask_id_workflowtaskv2"),
+            ondelete="SET NULL",
         ),
         sa.PrimaryKeyConstraint("id", name=op.f("pk_historyrun")),
     )
@@ -68,6 +70,7 @@ def upgrade() -> None:
             ["history_run_id"],
             ["historyrun.id"],
             name=op.f("fk_historyunit_history_run_id_historyrun"),
+            ondelete="CASCADE",
         ),
         sa.PrimaryKeyConstraint("id", name=op.f("pk_historyunit")),
     )
@@ -83,6 +86,7 @@ def upgrade() -> None:
             ["dataset_id"],
             ["datasetv2.id"],
             name=op.f("fk_historyimagecache_dataset_id_datasetv2"),
+            ondelete="CASCADE",
         ),
         sa.ForeignKeyConstraint(
             ["latest_history_unit_id"],
@@ -90,11 +94,13 @@ def upgrade() -> None:
             name=op.f(
                 "fk_historyimagecache_latest_history_unit_id_historyunit"
             ),
+            ondelete="CASCADE",
         ),
         sa.ForeignKeyConstraint(
             ["workflowtask_id"],
             ["workflowtaskv2.id"],
             name=op.f("fk_historyimagecache_workflowtask_id_workflowtaskv2"),
+            ondelete="CASCADE",
         ),
         sa.PrimaryKeyConstraint(
             "zarr_url",