+++ CHANGELOG.md
@@ -35,7 +35,8 @@ All this is from \#2287
     * Drop `FRACTAL_RUNNER_TASKS_INCLUDE_IMAGE` obsolete configuration variable (\#2359).
 * Development:
     * Add codespell to precommit (\#2358).
-
+* Testing:
+    * Use `fractal-task-tools` to build `fractal-tasks-mock` manifest (\#2374).
 
 # 2.13.1